docs: add better contributing guidelines
Have a link to the [Guidelines for the InterPlanetary JavaScript
Projects]. It is in a file called `CONTRIBUTING.md` so the GitHub
will automatically pick it up and display for first time contributors.

The `CONTRIBUTING.md` is intentionally without any repository specific
information so that it can be re-used on other repos as well.

[Guidelines for the InterPlanetary JavaScript Projects]: https://github.com/ipfs/community/blob/master/CONTRIBUTING_JS.md